chore(ui): one source for option colour tokens
resolveOptionColor existed twice: packages/workspace-ui/src/colors.ts and
apps/web/src/components/workspaces/property-utils.ts, with the same sixteen
presets and the same hex fallback in both. The token is pure presentation, so
it belongs in the kit; both former homes re-export it, and the export map gains
the subpath they now read it from.
